requirements.txt
# Audio Processing
librosa>=0.10.0
pyaudio>=0.2.13
soundfile>=0.12.0
numpy>=1.24.0

# Machine Learning
tensorflow>=2.13.0
transformers>=4.30.0
scikit-learn>=1.3.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data Processing
pandas>=2.0.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
